Avoid cloning the state everywhere
because the HTTP state was passed as reference to the parsing loop,
the new state had to be generated from cloned data.

Now, the ownership of the state is given to the parsing loop,
and the parsing loop just transfers ownership of the state data
from the previous state to the new one.

Preliminary tests show a gain of 30% in parsing speed