Do not reregister interest on every call to ready()
The previous code got interest from the protocol implementation
on each call to readable(), writable(), back_readable() and
back_writable(), and changed the socket interest accordingly
(ie do we want to read and/or write data on the front socket
and the back socket).

Those calls to EventLoop::reregister were quite expensive
(spending 10% of the profiled time).

The new implementation registers interest once, for read, write,
hup and error. It will then store the readiness state (did the
socket indicate that it is readable or writable?) and the
interest state (do we need to read or write) and match them
until we can't do anything else with this client, and return
from ready()