use unix sockets to send commands to a proxy
this is easier to use than a generic interface like what I tried
to do before. Using a socket to communicate with another thread
seems odd, but the Channel API fits nicely with the non blocking
event loop of the proxy, and the potentially blocking command
side. It also removes the confusion between mpsc's and mio's
channels.

This unix socket command channel will then make it easy to
finish the change to a multiprocess proxy