add a method to transform a CommandChannel to use other Tx and Rx types
this is useful at the boot of a worker, because it must receive its
configuration from the channel to start a proxy, then pass the channel
to the proxy that will receive messages.

this should be safe because the channels separate messages by null byte
boundaries. So you could receive a message of one type, then a null
byte, then a message of another type