Implement 100-Continue
the way it works, we recognize the header in the request and keep the
content length, but do not try to copy it. Once the server has answered,
we check if we're in 100-continue state, and if yes, then change the state
to be right after a request sent the headers. Once the body is sent to
the backend entirely, we reset the response state and wait for the real
server answer