handle config message aggregation
Now each config message is identified by a client message id,
which corresponds to the message id if it comes from a config client
(some config messages can be sent by the master itself, like the initial
state loading). Each message sent to a worker is identified
by a worker message id and the worker token, and there's a mapping
between (worker message id, worker token) and client message id.

Like this, for a simple config message (let's say CONFIG-0) sent to
2 workers, there would be (CONFIG-0, 0) and (CONFIG-0, 1).

But for an aggregation of message like LoadState (identified by
LOAD-STATE-X) with 2 config messages to load, there would be:
- (LOAD-STATE-X-0, 0)
- (LOAD-STATE-X-0, 1)
- (LOAD-STATE-X-1, 0)
- (LOAD-STATE-X-1, 1)

and the client would get the answer once all four of those worker
messages received an Ok or Error.

Right now this does not handle specific behaviour for some messages,
like aggregating metrics or storing the type of message (needed for messages
like SoftStop).