add a config message to upgrade a worker and transfer listener sockets
this uses the SCM_RIGHTS feature of unix sockets.

- the new message is UpgradeWorker(worker_id)
- it will first send (blocking) an Order::ReturnListenSockets to the old worker, then wait for the worker to send the sockets (only the HTTP listener is sent for now)
- once it gets the listener, it starts the new worker and sends the
listeners
- the worker waits for the listeners, then initializes the proxies

that means we can have a small window during which the listening sockets
are in transfer, and we're waiting for the new worker to boot to start
accepting connections again. This should be fast enough, and we still
won't lose any new connections