move back Session features in proxy::Server
This allows a much cleaner design, since there is no separation of client
slabs per proxy: they are all stored in the same one (this causes some
typing issues that could be resolved later).

This means no more token arithmetic, everything handled at the same
level, and possibly a slight perf boost, since we might have less
indirection.

This has been quickly tested, we will need to wait a bit before we're
sure it does not break everything.

The code for the OpenSSL based HTTPS proxy has not been fully ported
during this refactoring, thereare are a few changes to the traits to
perform.

There is still some inconsistencies with the TCP proxy (adding a new listener
does not work yet, as this message requires a special case)