flush the accept queue if we reached the maximum number of connections
before, we only stopped accepting new connections when we reached the
limit, so new connections could stay in the listen socket's queue until
there's enough room for a new client. Unfortunately, that means a new
connection could wait a long time and timeout (and occupy valuable
ressources in the meantime).

Since reaching the limit is supposed to be a rare case, and we want to
keep acceptable performance even in degenrate cases, this commit lets us
close all new pending connections when're under already heavy load, and
start accepting again when we have enough room.