Rework the main() function to use a Result oriented execution flow
The previous main function had duplicated code and was based on if /
else for execution flow based on the return of some functions.

Now, we (almost) only use the Result type to describe the flow and have
something easier to read or to maintain and update.

I also removed some duplicated code on the clap parsing and exported it
in a different file (src/cli.rs)