Proxys: warn / answer an order message error when there is no proxy to handle the message
This fixes two problems:
- first, if you try to add a frontend (http / tcp) with sozuctl, it will
timeout because we'd never answer. Now, it answers an error.
- second, no error message is shown when adding an http / tcp frontend
loaded from the config.toml file. This means that you can declare as
much as frontend you want and never see a message letting you know that
they weren't added. It may confuse users (and it consfused me).