do not use BackendConnection::Replace when retrying a backend connection
if the backend was down, when receiving the next event after trying to
connect the backend, we would send a `ReconnectBackend` message which
would result in the token used for that backend removed from the clients
slab.
The backend connection code used the presence of an app id in the HTTP
state machine to decide if it was for the first request of a connection,
or if we did a new request on the same connection.
In the ReconnectBackend case, we had the same app id, but the backend
was already removed from the clients slab, so we could not use the same
backend token.

This commit uses the presence of the backend_token field to decide if
it's a replace or new connection