Replace the metrics BufWriter with a LineWriter
with the BufWriter's implementation of fmt::Write, it could flush
the buffer right in the middle of a format string, so it would truncate
a metrics packet right at a format boundary.

Example: in "{}.backend.{},origin={},version={},app_id={},backend_id={}:{}|g\n"
it could flush right after "app_id=" and the resulting line would be
invalid.

In some small profiling sessions, the metrics part goes from 3% usage to
3.7% with this fix, so there should be a better solution that aggregates
multiple lines, but in the meantime we'll get all the metrics correctly