reintroduce timeouts
we employ the time wheels that were extracted from mio to mio-more then
mio-extras.

We set up a global timer, and for each new accepted connection, set up a
timeout of 60 seconds (hardcoded for now).
Each client session remembers the last time an event (any event on any
of its socket) was received. When the timeout triggers, we check the duration
from the last events, and if it's less than 60 seconds, we register a
new timeout. If it's more than 60 seconds, we close the session.
Whenever a session closes, we cancel all of its timeouts, to prevent
them from triggering for unrelated connections (since tokens are
reused).