Properly handle HUP events for back sockets
proxying sessions can get to a case where we get a HUP event for the
back socket (meaning it's closed), but we can still have some data
available, either in sozu's buffers, or in the kernel's buffers.
The first case was handled before by checking if the front socket was
writable, but in the second case we could have the front socket writable
nothing in sozu's buffers, but the kernel buffers still containing some
data.

In some tests we saw that when w get that event, the kernel buffers
could still store around 300kB of data