implement a zombie check
there can be bugs in the event loop implementation that result in some
session never receiving any new event, and forever holding slots in the
slab and buffers from the pool.

This checks regularly that there are no sessions that are too old, and
deletes them after logging debug info.
This allows sozu to reduce the impact of these memory leaks, while
detecting bugs and provide information for the fix in the logs