refactor the Readiness structure
the previous structure contained event and interest data for the front
and back socket.
With the new structure, we need one Readiness for a front socket, and
one for a back socket. This will be useful for HTTP/2 where we can have
multiple backend sockets in one session, and we will be able to move the
readiness information in the same structure as the socket itself