implement an accept queue
instead of accepting new sockets once we have room for new client
sessions (with the buffer pool usage, etc), we accept as many sockets as
possible, and put them in a LIFO queue, storing the date at which it was
accepted. We then go through that queue and create as many clients as
possible.
In normal load, the queue is always empty. Under considerable load, the
accept queue will fill up, and clients will be created in priority for
the most recent sockets (thanks to the LIFO queue), and sockets older
than 60s will be closed directly.
This allows us to keep wait times low even when under load. With the
previous code, when sozu was under load, we would avoid accepting new
connections. But those connections were already accepted by the kernel
(TCP handshake) and are holding kernel resources (some data can even be
already available). And if we take too much time accepting those, we
could get in a case where we parse that available data, send to backend,
get the response, but when we send the response to the client, the
client has already dropped the connection (timeout).
By closing sockets that are too old, we can signal early to the client
that we're overloaded, and they should try later.