replace HttpsFront with HttpFront
the fingerprint field was used to count how many frontends are linked to
a particular certificate, and have some refcounting layer to know if we
can remove a certificate (if some frontends still use it, we can't).
This makes the code needlessly complex. It is not an issue to have some
frontends using domains that are not handled by any certificate, it only
means no request will ever use those frontends.
Instead, we now make the certificates independent from their frontend,
and let the configuration layer handle the link between certificates and
frontends.
We rely on openssl and rustls refcounting for the certificates, and
replace HttpsFront with Httpfront everywhere