force session shutdown in more cases
shutting down or upgrading had issues with long running connections
(WebSockets or HTTP keep alive with regular requests). Those sessions
were never closing and preventing the shutdown of workers.

For websockets, we cannot know when the connection can be safely closed,
as it depends on the application's protocol, but websocket clients are
commonly designed to handle reconnections. So websocket connections will
be closed directly during shutdown.

For HTTP sessions with keep alive, the right way would be to add a "Connection:
Close" header on the next request to the session, and to the next response,
and after that response we can safely close. This is a bit difficult to
do for now, and I'd rather wait for the HTTP implementation rewrite for
this. In the meantime, it is allowed to close an idle connection (the
entire response was sent, waiting for the next request), so if a sssion
is in that state, we will close it. There is still an edge case for
sessions where any time we check its status, it is again in the middle
of a request or response