do not forget to close the listener sockets on the master process
when transmitting listener sockets from one worker to a new one, we
first transmit them to the master process, then from the master process
to the new worker. This means that the listener socket will be
duplicated in the master process as well. If we do not close them on the
master after transmitting to the new worker, the socket on the master's
side can receive new connections in its own accept queue, and they will
never reach the worker