insert Connection: close headers when shutting down the proxy
keep alive connections were able to continue running indefinitely and
prevent the proxy from shutting down. With this change, once we receive
the shutdown command, every HTTP session goes into "closing" mode, where
a "Connection: close" header is added when sending the request to the
backend, and the response to the client, then the session is closed when
the response is done.
Keep alive connections that do not send any requests will not be closed
immediately, but the timeout will stop them after some time.

The header is added directly instead of editing whatever value we might
have put in the Connection header before, so it is not completely
correct, but good enough for now