use the poule crate for the buffer pool
this gives us a pool that will not change its memory address (so it
could store buffers), but can increase its capacity.
This means sozu can start with minimal memory usage, but allocate as needed,
instead of preallocating all the buffers, slowing down at boot and causing
OOM kills during upgrades.

As a principle, we want to avoid allocations on the hot path, but
allocations on this pool will be very cheap:
- increasing capacity once in a while by asking for new memory pages
with mprotect (right now, we double capacity at each step, maybe smarter
strategies could be used)
- setting up a new buffer or resetting it means setting 2 usize elements
to zero