aggregate stored metrics
since metrics will be generated for each cluster and each backend, we
will store a lot of different points. Since this data is only meant for
quick debugging, we do not need to store long term data, and we do not
need very precise data.
So, to preerve space, we will regularly agregate data:
- all the new measures get aggregated in a point with the timestamp set
at the last second
- every minute, we aggregate the 60 previous measures in a point set at
one minute ago
- every hour, we aggregate the 60 previous "minute points" in a point
set at one hour ago, and we delete all of the points older than 24h

This introduces a lot of bookkeeping, so there will be runtime commands
to reset the database to an empty file, and activate or deactivate the
collection