send a status request to the new worker during an upgrade
the upgrade process was creating the new worker, then immediately
telling the old one to return the listen sockets, then sending them to
the new one. Unfortunately, if the configuration state was large, the
new worker could take some time loading it before receiving the listen
sockets and starting to accept new ones, so some sockets in queue were
timing out.

Now we send a status message (if Server is started with the
expect_initial_status option) and wait for the worker to answer before
telling the old worker to give back the listen sockets