store an instance of Proxy in sessions, handle close() in ready()
with the Proxy instance available in sessions, they can access the
sessions slab and the registry to close sockets without going all the
way up to the event loop.

Once the same is done for other SessionResult variants (CloseBackend,
ConnectBackend and ReConnecBackend), we will be able to remove methods
from the ProxySession trait and handle most of the socket management
from inside sessions, which will greatly simplify the code.

With this design though, we need to take extra care of checking that we
are not borrowing the session or the proxy multiple times (which would
make sozu panic). Once the transition from the previous design to this
one is done, we should be able to clean up the code and avoid most
borrowing issues