store cluster metrics in a simpler way, query them all
This commit removes what @Geal started to work on but didn't finish.
The cluster metrics are now stored without timestamps, which reduces
the strain on the memory and thus is cleared only everyday at midgnight.
They are stored in a simple binary tree map of the form
cluster_or_backend_id -> (key, value)

This commit removes the option to enable the timed collection of metrics.
Left to do: queries for cluster and backend metrics, providing their
ids and metric names.