retrieve cluster and backend metrics by ids
For each cluster, we display the metrics that belong to it
and the metrics of the attached backends.
This commit redefines WorkerMetrics served in ProxyResponseData,
with plenty of Options in it and its children strucs, to enable
partial answers: only proxy metrics, only certain clusters or
certain backends, or everything at once.